[buffer] get occupation vectors by rescaled offsetted points
This fixes the last case rt_p20

Check if the occupation vectors are short (length 1). This indicates a rounding issue.
If so, map again but use neighbouring cells.

Alas we have to do this, but still better than the former FP implementation.

Also, we map only points on offsetted borders now, and return if the map is empty,
to improve performance.