[index][strategies][test] Modify some of the header includes.
The purpose of this is to allow the compilation of the rtree without the need of including the geometry.hpp ond require only index/rtree.hpp or to make the order of includes irrelevant.
Furthermore, this is backward compatible.

Add missing variant include in the intersection tests.
Fix the unused typedefs in the pythagoras_point_box test.