[views] Fix compilation error in box_view caused by the change of a detail::assign_box_corners_oriented() body.
This function requires Range but was using operator[] which isn't a part of Range concept.
And box_view passed a raw pointer into this function.
So now a temporary range std::pair<> is created and passed, this might be changed if point_view stored boost:array<>.
It could be directly passed into the assign_box_corners_oriented.