[strategies] Add geographic side strategies based on azimuths calculation.
Three strategies added: andoyer, thomas and vincenty.
The strategies use inverse formulas for azimuths calculation.
Azimuths are compared to check the side.

The agnostic side_by_azimuth strategy was removed/replaced.
The code was moved into the geographic directory, into the details.
After some modifications it's used as a base of the geographic side strategies.