[test][util] Fix formatting errors and compatibility with other tests
Add a BOOST_GEOMETRY_TEST_USE_UTF macro to trigger the inclusion of the UTF. By default the test excecution monitor (used by the others test) is included but both cannot be included at the same time. This macro should help the transition to the UTF and will be removed when all tests will be based on UTF (remain to see if that is desirable).
Remove unused `using namespace boost::unit_test;`