[index][pack_create] Fix for iterators not returning true references.
Don't dereference iterators passed by the user in the expression where a
dereferenced value is passed into the translator/indexable-getter.

Without this fix if an iterator returns non-true reference (e.g. like
segment_iterator) and this reference is adapted to the Geometry concept
(like pointing_segment returned by segment_iterator) and the default
indexable<> getter is used (which returns a true reference), then a true
reference returned by it is a dangling reference corresponding to a
destroyed temporary object.