[relate][util] Take special care about NaN coordinates in relate for MultiLinestring.
In places where a check must be performed, if a Point is one of the
endpoints of a Linestring contained in a MultiLinestring, std::sort() and
std::equal_range() algorithms are used. With MSVC the assertion in
std::equal_range() fails if the elements cannot be reliably compared, i.e.
in the case when Points has NaN coordinates.

Add has_nan_coordinate() utility and use it in boundary_checker and
topology_check in relate() implementation.