[algorithms][overlay] add preprocessing step in overlay algorithm: before computing the overlay, compute the self turns of the two geometries and insert the points corresponding to any touch interior self turns they have;
This preprocessing step is important for robustness reasons: suppose we have
a geometry that has a touch interior self turn, the touch point T of which lies
in the interior of some segment AB of the geometry. During a set operation the
segment AB may be clipped to another segment CD containting T, where at least
one of C and D lies in the interior of AB. By construction T must be a touch
point that lies in the interior of CD, but numerically (i.e., since C and D
may be computed only approximately) T may be considered as not a point on CD,
which can then indicate that the computed geometry is invalid.

The solution proposed in this commit basically inserts T in the geometry that
has T as a self turns, and this enforces T to be present on CD when the set
operation is performed.