[strategies] Improve robustness of geodesic intersection strategy.
Reverse segments before analysing them if needed. Make sure the segments
are going from S to N. This way the intersection point calculated by
Sjoberg's intersection formula is always the same for equal but reversed
segments. The latitude is taken into account because this way the accuracy
of result's longitudes is greater.