[core][overlay] Improve/fix add_rings().
Support passing of area strategy into add_rings() and pass it from
overlay, buffer and dissolve.

Add invalid_output_exception and if preprocesor definition is
defined throw it if an intermediate result in add_rings() has negative
area which means that in non-cartesian CS too big polygon is the result.

If area of the intermediate result is 0 ignore it.