[srs] Improve handling of projection parameters.
Replace pj_param() taking param type in string with param name and
returning pvalue<T> with pj_param_X() returning type specified as X
directly.

Add pj_param_X() overloads checking if the parameter exists and assigning
it in one call in order to avoid two searches of parmeter list. Use this
function in various places when the assignment depends on the existance of
the parameter.

Add pj_mkparam() overload taking name and value as separate arguments to
avoid parsing string to extract them.

Remove i and f members of pvalue<T> which were only used to hold return
value in pj_param() and assigned to unnecessarily in pj_mkparam().

Comment out used member of pvalue<T> which is not used.