[index] Traverse nodes differently in iterative distance query.
In iterative distance query prioritize also based on level. This should
ensure that leafs are reached asap in case there are many internal nodes
with the same distance.

Replace apply_visitor with get() after manually checking level. This is
the first step to support weak nodes - non-variant nodes without an id
stored internally to distinguish between internal nodes and leafs.

Restore recursion in spatial query because it is faster and will stay at
least for now because during destruction the rtree is traversed as well
and manual stack could throw bad_alloc.